fix(ses): receive symlinks
