+++ packages/ses/functions/ses-receive/.npmrc
@@ -1,1 +1,1 @@
-save-exact = true
+../../../shared/.npmrc
